# myproject-web/src/main/resources/openapi/myproject-rest-api.yaml
...
openapi: 3.0.3
info:
  title: My Project
  description: My Project's REST API
  version: v1
servers:
  - url: 'http://localhost:8000/api'
...
paths:
...
  /staff:
    get:
      summary: Finds staff members
      operationId: findStaff
      tags:
        - staff
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StaffPage'
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      summary: Creates a new staff member
      operationId: createStaff
      tags:
        - staff
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StaffForm'
      responses:
        201:
          description: Created
        400:
          description: Bad request. The staff data is not valid.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
security:
  - basicAuth: []
components:
  schemas:
    Address:
      type: object
      properties:
        line_1:
          type: string
        line_2:
          type: string
        postal_code:
          type: string
        city:
          type: string
        country:
          type: string
        phone_number:
          type: string
    AddressForm:
      type: object
      properties:
        line_1:
          type: string
        line_2:
          type: string
        postal_code:
          type: string
        district:
          type: string
        city_id:
          type: integer
        phone_number:
          type: string
      required:
        - line_1
        - postal_code
        - city_id
...
    ErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
        error:
          type: string
        message:
          type: string
        path:
          type: string
          format: uri
...
    Page:
      type: object
      properties:
        page:
          type: integer
        number_of_elements:
          type: integer
        total_elements:
          type: integer
    Person:
      type: object
      properties:
        first_name:
          type: string
        last_name:
          type: string
    PersonForm:
      type: object
      properties:
        first_name:
          type: string
          minLength: 2
          maxLength: 45
        last_name:
          type: string
          minLength: 2
          maxLength: 45
      required:
        - first_name
        - last_name
    Staff:
      allOf:
        - $ref: '#/components/schemas/Person'
        - type: object
          properties:
            id:
              type: integer
            email:
              type: string
              format: email
            active:
              type: boolean
            address:
              $ref: '#/components/schemas/Address'
    StaffForm:
      allOf:
        - $ref: '#/components/schemas/PersonForm'
        - type: object
          properties:
            username:
              type: string
              minLength: 3
            email:
              type: string
              format: email
            password:
              type: string
              format: password
            address:
              $ref: '#/components/schemas/AddressForm'
            store_id:
              type: integer
      required:
        - first_name
        - last_name
        - username
        - email
    StaffPage:
      allOf:
        - $ref: '#/components/schemas/Page'
        - type: object
          properties:
            staff:
              type: array
              items:
                $ref: '#/components/schemas/Staff'
...
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
      description: Basic Authentication